chore: remove InflightNetworkNote
